Separating out individual SSA op tests into their own file
[ghstack-poisoned]
